Lowercase all the rules text except SELF, PARENT, and NAME_ references. Update list of keywords.
Turn on backtracking and disable the acyclic DFA (with k=1) to
fix lexing woes.
Add WALK token.
Move the NON token to the top grammar to gain precedence.
Add a hack to the NOT token to allow somewhat correct recognition of
"that isn't" (currently recognized as THAT_IS NOT instead of THAT ISNT).