Remove contraction tokens.
Instead of requiring separate cases for IS, IS NOT, and ISNT,
have NOT match the latter part of the contraction so that "isn't"
becomes IS NOT. Similarly for the other contractions 're, 've, and 's.

This last one unfortunately can also be a possessive, so split POSS
up into APOS_S and S_APOS, and then let poss and is_ match as
appropriate.

Also merge the plural forms of IS and HAS into the singular forms, as
we don't really need to have ( IS | ARE ) strewn everywhere.