Move properties case 4 up a level into subset.
By allowing subset to have an embedded list of properties and
restrictions, we cover what case 4 used to cover, eg. "a permanent you
control or suspended card you own". This allows us to axe case 4 for
great performance benefit.

To clear up ambiguities this introduces, we set greedy=true in two
places in the properties rule, and split the old case 1 into a new case
1 and 4. This silences the warnings ANTLR gives for it.

It is then possible, by similarly applying greedy=true in subset_list,
to use subset_list in place of pl_subset where it is necessary to match
a list that can contain both players and objects.

Fix the small handful of test cases this affects.