Better handle NAME_ refs mixed with SELF.
Previously, we'd handle "named Card Name, Card Name" by suggesting
both "Card Name, Card Name" and "Card Name", and then choosing the
latter because it was a real card name--unless one of "and" or "or"
followed this phrase, in which case we'd only suggest "Card Name, Card Name",
which incorrectly processed Noble Banneret.

All three existing cards now correctly preprocess into
"named NAME_Card_Name, SELF".

Further, if no suggestions are actual card names, we now choose the
first suggestion not present in all_names rather than the last. This
corrects a single token name with a comma in it.