Fix using PARENT where SELF is more correct.
Some cards grant (or take away from) themselves abilities under certain
conditions, and these abilities that reference themselves should say
"SELF" rather than "PARENT". A quick search indicates that all abilities
that should be PARENT are granted to an equipped or enchanted object, or
a created token. So to determine which it should be, we search backwards
for the first occurrence of "equipped", "enchanted", "create", or
"SELF". And throw in fortified for good measure.

Some abilities say "as long as SELF is enchanted," but these are not
counted as matching the word "enchanted" thanks to the comma. If there's
ever a card with "as long as SELF is enchanted or equipped", then this
will need revision.

Additionally, cut out reminder text first, rather than preprocessing it
before deleting it.