Handle some previously unnoticed cost items.
Fix the cost regex that extracts ability costs to allow single quotes
that are not the start or end of an ability.

Add parser rules for paying the mana cost(s) of a subset, and returning
a subset to hand.

Add a special rule for ref_obj_poss that includes ITS.

Remove unused S_APOS token.